proceeding towards solutions.

+++ solution.c
@@ -1,11 +1,11 @@
-int s1(int a,int b,int c)
+float s1(float a,float b,float c)
 {
-  int pr1= (-b)+sqrt(b*b-(4*a*c));
+  float pr1= (-b)+sqrt(b*b-(4*a*c));
   return pr1;
 }
 
-int s2(int a,int b,int c)
+float s2(float a,float b,float c)
 {
-  int pr2=(-b)-sqrt(b*b-(4*a*c));
+  float pr2=(-b)-sqrt(b*b-(4*a*c));
   return pr2;
 }
